// bot/game/enums/weapon.ts
export enum WeaponType {
    ONEHANDEDSWORD = "ONEHANDEDSWORD",
    TWOHANDEDSWORD = "TWOHANDEDSWORD",
    ONEHANDEDAXE = "ONEHANDEDAXE",
    TWOHANDEDAXE = "TWOHANDEDAXE",
    KNIFE = "KNIFE",
    BOW = "BOW",
    TWOHANDEDCROSSBOW = "TWOHANDEDCROSSBOW",
    ONEHANDEDCROSSBOW = "ONEHANDEDCROSSBOW",
    STAFF = "STAFF",
    UNARMED = "UNARMED"
}

export enum TwoHandedAxeName {
    BARRACKAXE = "Barrack Axe",
    LUMBERAXE = "Lumber Axe",
    MOGROKSLOSTAXE = "Mogrok's Lost Axe"
}

export enum KnifeName {
    BUTTERKNIFE = "Butter Knife",
    KRISBLADE = "Kris Blade",
    LETTEROPENER = "Letter Opener",
    THIEFBLADE = "Thief Blade",
    SHANK = "Shank"
}

export enum TwoHandedSwordName {
    TRAININGSWORD = "Training Sword",
    SCUFFEDBLADE = "Scuffed Blade",
    OLDCLAYMORE = "Old Claymore",
    LIONHEART = "Lion Heart"
}

export enum OneHandedAxeName {
    RUSTYAXE = "Rusty Axe",
    HATCHET = "Hatchet",
    LUMBERWICKER = "Lumber Wicker"
}

export enum OneHandedSwordName {
    RUSTYSWORD = "Rusty Sword",
    CRACKEDSABRE = "Cracked Sabre",
    COPPERBLADE = "Copper Blade",
    KNIGHTSSIDE = "Knight's Side"
}

export enum StaffName {
    WALKINGSTICK = "Walking Stick",
}

export enum BowName {
    TWIGBOW = "Twig Bow",
    CRACKEDOAKBOW = "Cracked Oak Bow",
    WOODSMANBOW = "Woodsman Bow",
    ELEVENSOLIDERBOW = "Eleven Solider Bow",
    PLAINSTROLLBOW = "Plains Troll Bow",
}